+++ scripts/genzhi18n.js
@@ -1,7 +1,17 @@
+const arguments = process.argv.slice(2);
+let name = 'zh-cn';
+if (arguments.length > 0) {
+    name = arguments[0];
+}
+
 const en = require('../out/i18n/en');
-const zhCn = require('../out/i18n/zh-cn');
+const zhCn = require('../out/i18n/' + name);
 const fs = require("fs");
 
 const result = { ...en, ...zhCn };
 
-fs.writeFileSync('./scripts/zh-cn.json', JSON.stringify(result, Object.keys(result).sort(), 4));
+fs.writeFileSync('./scripts/' + name + '.ts',
+    `import { __table } from './en';\r\n/*eslint sort-keys: "warn"*/\r\nconst table: Partial<typeof __table> = ` +
+    JSON.stringify(result, Object.keys(result).sort(), 4) +
+    `;\r\n\r\nexport = table;\r\n`
+    );